@@ -1,27 +1,29 @@
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using Microsoft.EntityFrameworkCore;
+using Sample.APM.EFCore;
 using Wing;
 
-namespace Sample.APM.EFCore
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Run();
-        }
+var builder = WebApplication.CreateBuilder(args);
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.UseStartup<Startup>();
-                }).AddWing(builder => builder.AddConsul());
-    }
-}
+builder.Host.AddWing(builder => builder.AddConsul());
+
+builder.Services.AddControllers();
+
+// Add services to the container.
+builder.Services.AddDbContext<EFCoreDemoContext>(options =>
+               options.UseSqlServer(builder.Configuration.GetConnectionString("Wing.Demo")));
+
+builder.Services.AddWing()
+       .AddPersistence()
+       .AddAPM(x => x.AddEFCore());
+
+var app = builder.Build();
+
+// Configure the HTTP request pipeline.
+
+app.UseHttpsRedirection();
+
+app.UseAuthorization();
+
+app.MapControllers();
+
+app.Run();
