@@ -1,27 +1,34 @@
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Configuration;
-using Microsoft.Extensions.Hosting;
-using Microsoft.Extensions.Logging;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
+using SqlSugar;
 using Wing;
 
-namespace Sample.APM.SqlSugar
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Run();
-        }
+var builder = WebApplication.CreateBuilder(args);
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-                .ConfigureWebHostDefaults(webBuilder =>
-                {
-                    webBuilder.UseStartup<Startup>();
-                }).AddWing(builder => builder.AddConsul());
-    }
-}
+builder.Host.AddWing(builder => builder.AddConsul());
+
+builder.Services.AddControllers();
+builder.Services.AddWing()
+                .AddPersistence()
+                .AddAPM(x => x.AddSqlSugar());
+builder.Services.AddScoped<ISqlSugarClient>(s =>
+{
+    SqlSugarClient sqlSugar = new SqlSugarClient(new ConnectionConfig()
+    {
+        DbType = DbType.SqlServer,
+        ConnectionString = builder.Configuration["ConnectionStrings:Wing.Demo"],
+        IsAutoCloseConnection = true,
+    },
+   db => db.AddWingAPM());
+    return sqlSugar;
+});
+
+var app = builder.Build();
+
+// Configure the HTTP request pipeline.
+
+app.UseHttpsRedirection();
+
+app.UseAuthorization();
+
+app.MapControllers();
+
+app.Run();
