@@ -1,21 +1,28 @@
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Hosting;
+using System.Security.Claims;
 using Wing;
 
-namespace Sample.AspNetCoreService2
-{
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Run();
-        }
+var builder = WebApplication.CreateBuilder(args);
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-             .ConfigureWebHostDefaults(webBuilder =>
-             {
-                 webBuilder.UseStartup<Startup>();
-             }).AddWing(builder=> builder.AddConsul());
-    }
-}
+builder.Services.AddControllers();
+builder.Services.AddHttpClient();
+builder.Host.AddWing(builder => builder.AddConsul());
+
+builder.Services.AddWing().AddJwt(context =>
+{
+    var user = context.User.Claims.Where(c => c.Type == ClaimTypes.Name).FirstOrDefault().Value;
+    return user == "byron";
+});
+
+var app = builder.Build();
+
+// Configure the HTTP request pipeline.
+
+app.UseHttpsRedirection();
+
+app.UseAuthentication();
+
+app.UseAuthorization();
+
+app.MapControllers();
+
+app.Run();
