@@ -1,21 +1,37 @@
-using Microsoft.AspNetCore.Hosting;
-using Microsoft.Extensions.Hosting;
+using FreeSql;
+using Sample.AspNetCoreService;
+using System.Security.Claims;
 using Wing;
+using Wing.ServiceProvider;
 
-namespace Sample.AspNetCoreService
+var builder = WebApplication.CreateBuilder(args);
+
+builder.Services.AddControllers();
+builder.Services.AddHttpClient();
+builder.Host.AddWing(builder => builder.AddConsul());
+
+var fsql = new FreeSqlBuilder()
+                              .UseConnectionString(DataType.SqlServer, builder.Configuration["ConnectionStrings:Sample.Wing"])
+                              .UseAutoSyncStructure(true) // �Զ�ͬ��ʵ��ṹ�����ݿ⣬FreeSql����ɨ����򼯣�ֻ��CRUDʱ�Ż����ɱ���
+                              .Build<SampleWingDbFlag>();
+builder.Services.AddSingleton(typeof(IFreeSql<SampleWingDbFlag>), serviceProvider => fsql);
+builder.Services.AddSingleton<ITracerService, TracerService>();
+builder.Services.AddWing().AddPersistence().AddEventBus().AddJwt(context =>
 {
-    public class Program
-    {
-        public static void Main(string[] args)
-        {
-            CreateHostBuilder(args).Build().Run();
-        }
+    var user = context.User.Claims.Where(c => c.Type == ClaimTypes.Name).FirstOrDefault().Value;
+    return user == "byron";
+}).AddAPM(x => x.AddFreeSql().Build(fsql));
 
-        public static IHostBuilder CreateHostBuilder(string[] args) =>
-            Host.CreateDefaultBuilder(args)
-             .ConfigureWebHostDefaults(webBuilder =>
-             {
-                 webBuilder.UseStartup<Startup>();
-             }).AddWing(builder=> builder.AddConsul());
-    }
-}
+var app = builder.Build();
+
+// Configure the HTTP request pipeline.
+
+app.UseHttpsRedirection();
+
+app.UseAuthentication();
+
+app.UseAuthorization();
+
+app.MapControllers();
+
+app.Run();
